The left most tile in the map can be viewed
Before this area was hidden by the GUI, but now by dynamically changing
the boundry of the camera that area previously hidden can now be viewed.

To do this a variable in the camera called GUI_LEFT was added that
contains the width of the left of the GUI, it was called this for
potential GUI_TOP, GUI_RIGHT, GUI_BOTTOM to be added should it become more
advanced.